compiler warns

+++ src/ScreenSMOnlineLogin.cpp
@@ -92,7 +92,7 @@ void ScreenSMOnlineLogin::ExportOptions( int iRow, const vector<PlayerNumber> &v
 
 			FOREACH_EnabledPlayer(pn) {
 				auto selection = m_pRows[0]->GetOneSelection(pn);
-				if (selection < m_pRows[0]->GetHandler()->m_Def.m_vsChoices.size() - 1) {
+				if (selection < static_cast<int>(m_pRows[0]->GetHandler()->m_Def.m_vsChoices.size()) - 1) {
 					ProfileManager::m_sDefaultLocalProfileID[pn].Set(vsProfiles[selection]);
 					typeUsername = false;
 				}
